requirements.txt
alabaster==0.7.16
ale-py==0.11.1
annotated-types==0.7.0
anyio==4.3.0
apturl==0.5.2
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
bcrypt==3.2.0
blinker==1.4
box2d-py==2.3.5
Brlapi==0.8.3
capstone==6.0.0a1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
colorama==0.4.4
colored-traceback==0.4.2
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dep-logic==0.2.0
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.20.1
duplicity==0.8.21
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fasteners==0.14.1
filelock==3.13.4
findpython==0.6.1
fiona==1.9.6
fonttools==4.51.0
fsspec==2025.3.2
future==0.18.2
geopandas==0.14.3
gymnasium==1.1.1
h11==0.14.0
hishel==0.0.26
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
iniconfig==2.0.0
installer==0.7.0
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.7
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.46.0
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.11.0.86
osmnx==1.9.2
packaging==24.0
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pbs-installer==2024.4.1
pdm==2.15.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.5.0
plumbum==1.9.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pwn==1.0
pwntools==4.13.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
pyelftools==0.31
pygame==2.6.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==2.4.7
pyproj==3.6.1
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
pytest==8.2.2
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.1
reportlab==3.6.8
requests==2.31.0
resolvelib==1.0.1
rich==13.7.1
ROPGadget==7.5
rpyc==6.0.1
scikit-learn==1.7.2
scipy==1.13.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
socksio==1.0.0
sortedcontainers==2.4.0
Sphinx==7.4.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
swig==4.3.1
sympy==1.14.0
systemd-python==234
threadpoolctl==3.4.0
tomli==2.0.1
tomlkit==0.12.4
torch==2.9.0+cu128
torchaudio==2.9.0+cu128
torchvision==0.24.0+cu128
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.0
truststore==0.8.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
umap-learn==0.5.9.post2
unattended-upgrades==0.1
unearth==0.15.2
unicorn==2.1.1
unix-ar==0.2.1
urllib3==1.26.5
usb-creator==0.3.7
uv==0.6.3
vbll==0.4.9
virtualenv==20.25.3
wadllib==1.3.6
wcwidth==0.2.13
widgetsnbextension==4.0.14
xdg==5
xkit==0.0.0
zipp==1.0.0
zstandard==0.22.0
